+++ python/egglog/exp/array_api_numba.py
@@ -9,7 +9,7 @@ from egglog import *
 from egglog.exp.array_api import *
 
 array_api_numba_ruleset = ruleset()
-array_api_numba_schedule = (array_api_ruleset + array_api_numba_ruleset).saturate()
+array_api_numba_schedule = (array_api_combined_ruleset | array_api_numba_ruleset).saturate()
 # For these rules, we not only wanna rewrite, we also want to subsume the original expression,
 # so that the rewritten one is used, even if the original one is simpler.
 
@@ -40,24 +40,19 @@ def _std(y: NDArray, x: NDArray, i: Int):
 
 
 # rewrite unique_counts to count each value one by one, since numba doesn't support np.unique(..., return_counts=True)
-@function
-def count_values(x: NDArray, values: NDArray) -> TupleValue:
+@function(ruleset=array_api_numba_ruleset, subsume=True)
+def count_values(x: NDArray, values: TupleValue) -> TupleValue:
     """
-    Returns a tuple of the count of each of the values in the array.
+    Returns a tuple of the count of each of the values in the array as values
     """
+    return TupleValue(values.length(), lambda i: sum(x == values[i]).to_value())
 
 
 @array_api_numba_ruleset.register
 def _unique_counts(x: NDArray, c: NDArray, tv: TupleValue, v: Value):
     return [
         # The unique counts are the count of all the unique values
-        rewrite(unique_counts(x)[Int(1)], subsume=True).to(NDArray.vector(count_values(x, unique_values(x)))),
-        rewrite(count_values(x, NDArray.vector(TupleValue(v) + tv)), subsume=True).to(
-            TupleValue(sum(x == NDArray.scalar(v)).to_value()) + count_values(x, NDArray.vector(tv))
-        ),
-        rewrite(count_values(x, NDArray.vector(TupleValue(v))), subsume=True).to(
-            TupleValue(sum(x == NDArray.scalar(v)).to_value()),
-        ),
+        rewrite(unique_counts(x)[1], subsume=True).to(NDArray.vector(count_values(x, unique_values(x).to_values()))),
     ]