+++ python/tests/__snapshots__/test_array_api/test_jit[add][expr].py
@@ -1,0 +1,1 @@
+NDArray.var("x") + NDArray.var("y")