+++ python/tests/test_array_api.py
@@ -1,5 +1,7 @@
+# mypy: disable-error-code="empty-body"
 import ast
 from collections.abc import Callable
+from itertools import product
 from pathlib import Path
 from typing import Any
 
@@ -10,70 +12,181 @@ from sklearn.discriminant_analysis import LinearDiscriminantAnalysis
 
 from egglog.egraph import set_current_ruleset
 from egglog.exp.array_api import *
+from egglog.exp.array_api import NDArray
 from egglog.exp.array_api_jit import jit
 from egglog.exp.array_api_loopnest import *
 from egglog.exp.array_api_numba import array_api_numba_schedule
 from egglog.exp.array_api_program_gen import *
+from egglog.exp.program_gen import Program
+
+some_shape = constant("some_shape", TupleInt)
+some_dtype = constant("some_dtype", DType)
+some_index = constant("some_index", TupleInt)
+some_length = constant("some_length", Int)
+some_value = constant("some_value", Value)
+some_int_index = constant("some_int_index", Int)
 
 
-def test_simplify_any_unique():
-    X = NDArray.var("X")
-    res = (
-        any(
-            (astype(unique_counts(X)[Int(1)], DType.float64) / NDArray.scalar(Value.float(Float(150.0))))
-            < NDArray.scalar(Value.int(Int(0)))
+@function(ruleset=array_api_ruleset)
+def is_even(x: Int) -> Boolean:
+    return x % 2 == 0
+
+
+class TestTupleValue:
+    def test_includes(self):
+        x = TupleValue.EMPTY.append(Value.bool(FALSE))
+        check_eq(x.contains(Value.bool(FALSE)), TRUE, array_api_schedule)
+        check_eq(x.contains(Value.bool(TRUE)), FALSE, array_api_schedule)
+
+
+class TestTupleInt:
+    def test_cons_to_vec(self):
+        check_eq(
+            TupleInt.EMPTY.append(2),
+            TupleInt.from_vec(Vec(Int(2))),
+            array_api_schedule,
+            add_second=False,
         )
-        .to_value()
-        .to_bool
-    )
 
-    egraph = EGraph()
-    egraph.register(res)
-    egraph.run(array_api_schedule)
-    egraph.check(eq(res).to(FALSE))
+    def test_vec_to_cons(self):
+        check_eq(
+            TupleInt.from_vec(Vec(Int(1), Int(2))),
+            TupleInt.EMPTY.append(1).append(2),
+            array_api_schedule,
+            add_second=False,
+        )
+
+    def test_indexing_cons(self):
+        check_eq(TupleInt.EMPTY.append(1).append(2)[Int(0)], Int(1), array_api_schedule)
+        check_eq(TupleInt.EMPTY.append(1).append(2)[Int(1)], Int(2), array_api_schedule)
+
+    def test_length_cons(self):
+        check_eq(TupleInt.EMPTY.append(1).append(2).length(), Int(2), array_api_schedule)
+
+    def test_fn_to_cons(self):
+        check_eq(TupleInt(2, lambda i: i), TupleInt.EMPTY.append(0).append(1), array_api_schedule, add_second=False)
+
+    def test_range_length(self):
+        check_eq(TupleInt.range(some_length).length(), some_length, array_api_schedule)
+
+    def test_range_index(self):
+        check_eq(
+            TupleInt.range(some_length)[some_int_index], check_index(some_length, some_int_index), array_api_schedule
+        )
+
+    def test_not_contains_example(self):
+        check_eq(TupleInt.from_vec(Vec(Int(0), Int(1))).contains(Int(3)), FALSE, array_api_schedule)
+
+    def test_contains_example(self):
+        check_eq(TupleInt.from_vec(Vec(Int(0), Int(3))).contains(Int(3)), TRUE, array_api_schedule)
+
+    def test_filter_append(self):
+        check_eq(
+            TupleInt.EMPTY.append(1).append(2).filter(is_even),
+            TupleInt.EMPTY.append(2),
+            array_api_schedule,
+            add_second=False,
+        )
+
+    def test_filter_range(self):
+        check_eq(TupleInt.range(4).filter(is_even), TupleInt.from_vec(Vec(Int(0), Int(2))), array_api_schedule)
+
+    def test_filter_lambda_length(self):
+        with set_current_ruleset(array_api_ruleset):
+            x = TupleInt.range(5).filter(lambda i: i < 2).length()
+        check_eq(x, Int(2), array_api_schedule)
 
 
-def test_tuple_value_includes():
-    x = TupleValue(Value.bool(FALSE))
-    should_be_true = x.includes(Value.bool(FALSE))
-    should_be_false = x.includes(Value.bool(TRUE))
-    egraph = EGraph()
-    egraph.register(should_be_true)
-    egraph.register(should_be_false)
-    egraph.run(array_api_schedule)
-    egraph.check(eq(should_be_true).to(TRUE))
-    egraph.check(eq(should_be_false).to(FALSE))
+@function
+def some_array_idx_fn(x: TupleInt) -> Value: ...
 
 
-def test_reshape_index():
-    # Verify that it doesn't expand forever
-    x = NDArray.var("x")
-    new_shape = TupleInt.single(Int(-1))
-    res = reshape(x, new_shape).index(TupleInt.single(Int(1)) + TupleInt.single(Int(2)))
-    egraph = EGraph()
-    egraph.register(res)
-    egraph.run(array_api_schedule)
-    equiv_expr = egraph.extract_multiple(res, 10)
-    assert len(equiv_expr) < 10
+class TestNDArray:
+    def test_index(self):
+        x = NDArray(some_shape, some_dtype, some_array_idx_fn)
+        check_eq(x.index(some_index), some_array_idx_fn(some_index), array_api_schedule)
+
+    def test_shape(self):
+        x = NDArray(some_shape, some_dtype, some_array_idx_fn)
+        check_eq(x.shape, some_shape, array_api_schedule)
+
+    def test_simplify_any_unique(self):
+        res = (
+            any(
+                (
+                    astype(unique_counts(NDArray.var("X"))[Int(1)], DType.float64)
+                    / NDArray.scalar(Value.float(Float(150.0)))
+                )
+                < NDArray.scalar(Value.int(Int(0)))
+            )
+            .to_value()
+            .to_bool
+        )
+        check_eq(res, FALSE, array_api_schedule)
+
+    def test_reshape_index(self):
+        # Verify that it doesn't expand forever
+        x = NDArray.var("x")
+        new_shape = TupleInt.single(Int(-1))
+        res = reshape(x, new_shape).index(TupleInt.single(Int(1)) + TupleInt.single(Int(2)))
+        egraph = EGraph()
+        egraph.register(res)
+        egraph.run(array_api_schedule)
+        equiv_expr = egraph.extract_multiple(res, 10)
+        assert len(equiv_expr) < 10
+
+    def test_reshape_vec_noop(self):
+        x = NDArray.var("x")
+        assume_shape(x, TupleInt.single(Int(5)))
+        res = reshape(x, TupleInt.single(Int(-1)))
+        egraph = EGraph()
+        egraph.register(res)
+        egraph.run(array_api_schedule)
+        equiv_expr = egraph.extract_multiple(res, 10)
+
+        assert len(equiv_expr) == 2
+        egraph.check(eq(res).to(x))
 
 
-def test_reshape_vec_noop():
-    x = NDArray.var("x")
-    assume_shape(x, TupleInt.single(Int(5)))
-    res = reshape(x, TupleInt.single(Int(-1)))
-    egraph = EGraph()
-    egraph.register(res)
-    egraph.run(array_api_schedule)
-    equiv_expr = egraph.extract_multiple(res, 10)
-
-    assert len(equiv_expr) == 2
-    egraph.check(eq(res).to(x))
+@function
+def some_tuple_tuple_int_idx_fn(x: Int) -> TupleInt: ...
 
 
-def test_filter():
-    with set_current_ruleset(array_api_ruleset):
-        x = TupleInt.range(5).filter(lambda i: i < 2).length()
-    check_eq(x, Int(2), array_api_schedule)
+@function
+def some_tuple_tuple_int_reduce_value_fn(carry: Value, x: TupleInt) -> Value: ...
+
+
+class TestTupleTupleInt:
+    def test_reduce_value_zero(self):
+        x = TupleTupleInt(0, some_tuple_tuple_int_idx_fn)
+        check_eq(x.foldl_value(some_tuple_tuple_int_reduce_value_fn, some_value), some_value, array_api_schedule)
+
+    def test_reduce_value_one(self):
+        x = TupleTupleInt(1, some_tuple_tuple_int_idx_fn)
+        check_eq(
+            x.foldl_value(some_tuple_tuple_int_reduce_value_fn, some_value),
+            some_tuple_tuple_int_reduce_value_fn(some_value, some_tuple_tuple_int_idx_fn(Int(0))),
+            array_api_schedule,
+        )
+
+    def test_product_example(self):
+        """
+        From Python docs:
+
+        product('ABCD', 'xy') → Ax Ay Bx By Cx Cy Dx Dy
+
+        aka product((0, 1, 2, 3), (4, 5)) ==
+        """
+        # TODO: Increase size, but for now check doesnt terminate at larger sizes for some reason
+        # input = ((0, 1, 2, 3), (4, 5))
+        input = ((0, 1), (4, 5))
+        expected_output = tuple(product(*input))
+        check_eq(
+            convert(input, TupleTupleInt).product(),
+            convert(expected_output, TupleTupleInt),
+            add_second=False,
+            schedule=array_api_schedule,
+        )
 
 
 @function(ruleset=array_api_ruleset, subsume=True)
@@ -89,77 +202,59 @@ def linalg_norm(X: NDArray, axis: TupleIntLike) -> NDArray:
         lambda k: LoopNestAPI.from_tuple(reduce_axis)
         .unwrap()
         .indices()
-        .reduce_value(lambda carry, i: carry + ((x := X.index(i + k)).conj() * x).real(), init=0.0)
+        .foldl_value(lambda carry, i: carry + ((x := X.index(i + k)).conj() * x).real(), init=0.0)
         .sqrt(),
     )
 
 
-X = NDArray.var("X")
-assume_shape(X, (3, 2, 3, 4))
-val = linalg_norm(X, (0, 1))
-i = constant("i", Int)
-j = constant("j", Int)
-idxed = val.index((i, j))
+@function(ruleset=array_api_ruleset, subsume=True)
+def linalg_norm_v2(X: NDArray, axis: TupleIntLike) -> NDArray:
+    return NDArray(
+        X.shape.deselect(axis),
+        X.dtype,
+        lambda k: ndindex(X.shape.select(axis))
+        .foldl_value(lambda carry, i: carry + ((x := X.index(i + k)).conj() * x).real(), init=0.0)
+        .sqrt(),
+    )
 
-egraph = EGraph(save_egglog_string=True)
-idxed = egraph.let("idxed", idxed)
-# egraph.run(array_api_ruleset * 43)
-# egraph.run(array_api_schedule)
 
-i = 0
-
-last = ""
-while True:
-    res = egraph.run(array_api_ruleset)
-    if not res.updated:
-        break
-    i += 1
-    #     if i < 43:
-    #         continue
-    # print([r for r, n in res.num_matches_per_rule.items() if n])
-    #     for r, n in res.num_matches_per_rule.items():
-    #         if n > 0:
-    #             print(r, n)
-    new = str(egraph.extract(idxed))
-    print(i)
-    if new == last:
-        continue
-    print(new, "\n\n")
-    last = new
-# try:
-# res = ""
-# n_same = 0
-# i = 0
-# while True:
-#     print(i, n_same)
-#     i += 1
-#     egraph.run(array_api_ruleset)
-#     new_res = str(egraph.extract(idxed))
-#     if new_res == res:
-#         n_same += 1
-#     else:
-#         n_same = 0
-#     if n_same == 15:
-#         break
-#     res = new_res
-# print(n_same)
-# print(res)
-# egraph.display(split_primitive_outputs=True, n_inline_leaves=3)
-# except:
-#     Path("tmp.egg").write_text(egraph.as_egglog_string)
-# egraph.run(array_api_schedule)
-print(egraph.extract(idxed))
+def linalg_val(linalg_fn: Callable[[NDArray, TupleIntLike], NDArray]) -> NDArray:
+    X = NDArray.var("X")
+    assume_shape(X, (3, 2, 3, 4))
+    return linalg_fn(X, (0, 1))
 
 
 class TestLoopNest:
-    def test_shape(self):
-        X = NDArray.var("X")
-        assume_shape(X, (3, 2, 3, 4))
-        val = linalg_norm(X, (0, 1))
+    @pytest.mark.parametrize("linalg_fn", [linalg_norm, linalg_norm_v2])
+    def test_shape(self, linalg_fn):
+        check_eq(linalg_val(linalg_fn).shape, TupleInt.from_vec(Vec(Int(3), Int(4))), array_api_schedule)
 
-        check_eq(val.shape.length(), Int(2), array_api_schedule)
-        check_eq(val.shape[0], Int(3), array_api_schedule)
-        check_eq(val.shape[1], Int(4), array_api_schedule)
+    @pytest.mark.parametrize("linalg_fn", [linalg_norm, linalg_norm_v2])
+    def test_index(self, linalg_fn):
+        i = constant("i", Int)
+        j = constant("j", Int)
+        idxed = linalg_val(linalg_fn).index((i, j))
+        _NDArray_1 = NDArray.var("X")
+        _Value_1 = _NDArray_1.index(TupleInt.from_vec(Vec[Int](Int(0), Int(0), i, j)))
+        _Value_2 = _NDArray_1.index(TupleInt.from_vec(Vec[Int](Int(0), Int(1), i, j)))
+        _Value_3 = _NDArray_1.index(TupleInt.from_vec(Vec[Int](Int(1), Int(0), i, j)))
+        _Value_4 = _NDArray_1.index(TupleInt.from_vec(Vec[Int](Int(1), Int(1), i, j)))
+        _Value_5 = _NDArray_1.index(TupleInt.from_vec(Vec[Int](Int(2), Int(0), i, j)))
+        _Value_6 = _NDArray_1.index(TupleInt.from_vec(Vec[Int](Int(2), Int(1), i, j)))
+        res = (
+            (
+                (
+                    (
+                        ((_Value_1.conj() * _Value_1).real() + (_Value_2.conj() * _Value_2).real())
+                        + (_Value_3.conj() * _Value_3).real()
+                    )
+                    + (_Value_4.conj() * _Value_4).real()
+                )
+                + (_Value_5.conj() * _Value_5).real()
+            )
+            + (_Value_6.conj() * _Value_6).real()
+        ).sqrt()
+        check_eq(idxed, res, array_api_schedule)
 
 
 # This test happens in different steps. Each will be benchmarked and saved as a snapshot.
@@ -169,7 +264,7 @@ class TestLoopNest:
 def run_lda(x, y):
     with config_context(array_api_dispatch=True):
         lda = LinearDiscriminantAnalysis(n_components=2)
-        return lda.fit(x, y).transform(x)
+        return lda.fit_transform(x, y)
 
 
 iris = datasets.load_iris()
@@ -201,7 +296,12 @@ def load_source(fn_program: EvalProgram, egraph: EGraph):
     egraph.register(fn_program)
     egraph.run(array_api_program_gen_schedule)
     # dp the needed pieces in here for benchmarking
-    return egraph.eval(egraph.extract(fn_program.py_object))
+    try:
+        return egraph.eval(egraph.extract(fn_program.py_object))
+    except Exception as err:
+        err.add_note(f"Failed to compile the program into a string: \n\n{egraph.extract(fn_program)}")
+        egraph.display(split_primitive_outputs=True, n_inline_leaves=3, split_functions=[Program])
+        raise
 
 
 def lda(X, y):
@@ -219,6 +319,40 @@ def simplify_lda(egraph: EGraph, expr: NDArray) -> NDArray:
     egraph.register(expr)
     egraph.run(array_api_numba_schedule)
     return egraph.extract(expr)
+
+
+@pytest.mark.parametrize(
+    "program",
+    [
+        pytest.param(tuple_value_program(TupleInt.from_vec((1, 2))), id="tuple"),
+    ],
+)
+def test_program_compile(program: Program, snapshot_py):
+    # simplify first to do any pre-conversion
+    egraph = EGraph()
+    egraph.register(program)
+    egraph.run(array_api_numba_schedule)
+    simplified_program = egraph.extract(program)
+    assert str(simplified_program) == snapshot_py(name="expr")
+    egraph = EGraph()
+    egraph.register(simplified_program.compile())
+    egraph.run(array_api_program_gen_schedule)
+    statements = egraph.eval(simplified_program.statements)
+    expr = egraph.eval(simplified_program.expr)
+    assert "\n".join([*statements.split("\n"), expr]) == snapshot_py(name="code")
+
+
+@pytest.mark.parametrize(
+    "program",
+    [
+        pytest.param(lambda x, y: x + y, id="add"),
+        pytest.param(lambda x, y: x[(x.shape + TupleInt.from_vec((1, 2)))[100]], id="tuple"),
+    ],
+)
+def test_jit(program, snapshot_py):
+    jitted = jit(program)
+    assert str(jitted.expr) == snapshot_py(name="expr")
+    assert jitted.statements == snapshot_py(name="code")
 
 
 @pytest.mark.benchmark(min_rounds=3)
@@ -262,35 +396,39 @@ class TestLDA:
         assert egraph.eval(fn_program.statements) == snapshot_py
 
     @pytest.mark.parametrize(
-        "fn",
+        "fn_thunk",
         [
-            pytest.param(LinearDiscriminantAnalysis(n_components=2).fit_transform, id="base"),
-            pytest.param(run_lda, id="array_api"),
-            pytest.param(_load_py_snapshot(test_source_optimized, "__fn"), id="array_api-optimized"),
-            pytest.param(numba.njit(_load_py_snapshot(test_source_optimized, "__fn")), id="array_api-optimized-numba"),
-            pytest.param(jit(lda), id="array_api-jit"),
+            pytest.param(lambda: LinearDiscriminantAnalysis(n_components=2).fit_transform, id="base"),
+            pytest.param(lambda: run_lda, id="array_api"),
+            pytest.param(lambda: _load_py_snapshot(TestLDA.test_source_optimized, "__fn"), id="array_api-optimized"),
+            pytest.param(
+                lambda: numba.njit(_load_py_snapshot(TestLDA.test_source_optimized, "__fn")),
+                id="array_api-optimized-numba",
+            ),
+            pytest.param(lambda: jit(lda), id="array_api-jit"),
         ],
     )
-    def test_execution(self, fn, benchmark):
+    def test_execution(self, fn_thunk, benchmark):
+        fn = fn_thunk()
         # warmup once for numba
-        assert np.allclose(run_lda(X_np, y_np), fn(X_np, y_np))
+        assert np.allclose(run_lda(X_np, y_np), fn(X_np, y_np), rtol=1e-03)
         benchmark(fn, X_np, y_np)
 
 
 # if calling as script, print out egglog source for test
 # similar to jit, but don't include pyobject parts so it works in vanilla egglog
-# if __name__ == "__main__":
-#     print("Generating egglog source for test")
-#     egraph = EGraph(save_egglog_string=True)
-#     X_ = NDArray.var("X")
-#     y_ = NDArray.var("y")
-#     with egraph:
-#         expr = lda(X_, y_)
-#     optimized_expr = egraph.simplify(expr, array_api_numba_schedule)
-#     fn_program = ndarray_function_two_program(optimized_expr, X_, y_)
-#     egraph.register(fn_program.compile())
-#     egraph.run(array_api_program_gen_ruleset.saturate() + program_gen_ruleset.saturate())
-#     egraph.extract(fn_program.statements)
-#     name = "python.egg"
-#     print("Saving to", name)
-#     Path(name).write_text(egraph.as_egglog_string)
+if __name__ == "__main__":
+    print("Generating egglog source for test")
+    egraph = EGraph(save_egglog_string=True)
+    X_ = NDArray.var("X")
+    y_ = NDArray.var("y")
+    with egraph:
+        expr = lda(X_, y_)
+    optimized_expr = egraph.simplify(expr, array_api_numba_schedule)
+    fn_program = ndarray_function_two_program(optimized_expr, X_, y_)
+    egraph.register(fn_program.compile())
+    egraph.run(array_api_program_gen_ruleset.saturate() + program_gen_ruleset.saturate())
+    egraph.extract(fn_program.statements)
+    name = "python.egg"
+    print("Saving to", name)
+    Path(name).write_text(egraph.as_egglog_string)